+++ app/models/facilities_management/rm6378/journey/choose_locations.rb
@@ -7,6 +7,11 @@ module FacilitiesManagement
       validates :region_codes, length: { minimum: 1 }
 
       attribute :annual_contract_value, Numeric
+      attribute :contract_start_date_dd
+      attribute :contract_start_date_mm
+      attribute :contract_start_date_yyyy
+      attribute :estimated_contract_duration, Numeric
+      attribute :private_finance_initiative
 
       def next_step_class
         Journey::AnnualContractValue