+++ spec/models/facilities_management/rm6378/journey/choose_locations_spec.rb
@@ -30,13 +30,13 @@ RSpec.describe FacilitiesManagement::RM6378::Journey::ChooseLocations do
 
   describe '.permit_list' do
     it 'returns a list of the permitted attributes' do
-      expect(described_class.permit_list).to eq [:annual_contract_value, { region_codes: [] }]
+      expect(described_class.permit_list).to eq [:annual_contract_value, :contract_start_date_dd, :contract_start_date_mm, :contract_start_date_yyyy, :estimated_contract_duration, :private_finance_initiative, { region_codes: [] }]
     end
   end
 
   describe '.permitted_keys' do
     it 'returns a list of the permitted keys' do
-      expect(described_class.permitted_keys).to eq %i[region_codes annual_contract_value]
+      expect(described_class.permitted_keys).to eq %i[region_codes annual_contract_value contract_start_date_dd contract_start_date_mm contract_start_date_yyyy estimated_contract_duration private_finance_initiative]
     end
   end