+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add flag to show GCA branding ([PR 5153](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5153))
+
 ### Changed
 
 - Update the service descriptions for FM RM6378 ([PR 5123](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5123))
 - Update GOV.UK Frontend and CCS Frontend ([PR 5126](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5126))
+- Add new FM RM6378 question answers to summary ([PR 5146](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5146))
 
 ## [7.5.0] - 2025-12-09